NewsProject/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django
django-cors-headers==4.6.0
django-ninja==1.3.0
h11==0.14.0
idna==3.10
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
sniffio==1.3.1
sqlparse==0.5.2
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
